Move typedef to new file
